delete API

+++ backend/dmu-datastore/utils/ddsutils.py
@@ -32,32 +32,55 @@ class DDSUtils:
 
     # uploading file to blob storage
     def upload_file_to_azure_blob(self, file_path, file_name, folder):
-        blob_file_name = folder + "/" + file_name
-        blob_service_client = BlobServiceClient.from_connection_string(azure_connection_string)
-        blob_client = blob_service_client.get_blob_client(container=azure_container_name, blob=blob_file_name)
-        log.info(f'{folder} | Pushing {file_path} to azure at {blob_file_name} on a new fork......')
-        persister = threading.Thread(target=self.upload_file, args=(blob_client, file_path, folder))
-        persister.start()
-        '''
-        log.info(f'Pushing {file_path} to azure at {blob_file_name} synchronously....')
-        self.upload_file(blob_client, file_path)
-        '''
-        return f'{azure_link_prefix}{blob_file_name}'
+        try:
+            blob_file_name = folder + "/" + file_name
+            blob_service_client = BlobServiceClient.from_connection_string(azure_connection_string)
+            blob_client = blob_service_client.get_blob_client(container=azure_container_name, blob=blob_file_name)
+            log.info(f'{folder} | Pushing {file_path} to azure at {blob_file_name} on a new fork......')
+            blob_uploader = threading.Thread(target=self.upload_file, args=(blob_client, file_path, folder))
+            blob_uploader.start()
+            '''
+            log.info(f'Pushing {file_path} to azure at {blob_file_name} synchronously....')
+            self.upload_file(blob_client, file_path)
+            '''
+            return f'{azure_link_prefix}{blob_file_name}'
+        except Exception as e:
+            log.exception(f'Exception while getting Blob Client to azure blob storage: {e}', e)
+            self.update_upload_status(folder, file_path, True, False)
+            return None
 
     def upload_file(self, blob_client, file_path, upload_id):
         try:
             with open(file_path, "rb") as data:
                 blob_client.upload_blob(data, overwrite=True)
-            dds_repo.update_dds_metadata({"uploadId": upload_id}, {"uploadStatus": "Completed",
-                                                                   "lastUpdatedTimestamp": eval(
-                                                                       str(time.time()).replace('.', '')[0:13])})
-            os.remove(file_path)
+            self.update_upload_status(upload_id, file_path, False, True)
         except Exception as e:
             log.exception(f'Exception while pushing to azure blob storage: {e}', e)
-            dds_repo.update_dds_metadata({"uploadId": upload_id}, {"uploadStatus": "Failed",
-                                                                   "lastUpdatedTimestamp": eval(
-                                                                       str(time.time()).replace('.', '')[0:13])})
+            self.update_upload_status(upload_id, file_path, True, True)
 
     def camel_case(self, str):
         s = sub(r"(_|-)+", " ", str).title().replace(" ", "")
         return ''.join([s[0].lower(), s[1:]])
+
+    def update_upload_status(self, upload_id, file_path, upload_failed, after_acceptance):
+        log.info(f"{upload_id} | Updating Upload status for: {file_path}")
+        if upload_failed:
+            if after_acceptance:
+                if '___metadata___' in file_path:
+                    update_query = {"uploadStatus": "Failed",
+                                    "lastUpdatedTimestamp": eval(str(time.time()).replace('.', '')[0:13]),
+                                    "metadata_local": file_path}
+                else:
+                    update_query = {"uploadStatus": "Failed",
+                                    "lastUpdatedTimestamp": eval(str(time.time()).replace('.', '')[0:13]),
+                                    "media_local": file_path}
+            else:
+                update_query = {"uploadStatus": "Failed",
+                                "lastUpdatedTimestamp": eval(str(time.time()).replace('.', '')[0:13])}
+            dds_repo.update_dds_metadata({"uploadId": upload_id}, update_query)
+        else:
+            dds_repo.update_dds_metadata({"uploadId": upload_id}, {"uploadStatus": "Completed",
+                                                                   "lastUpdatedTimestamp": eval(
+                                                                       str(time.time()).replace('.', '')[0:13])})
+            #dds_repo.update_dds_metadata({"uploadId": upload_id}, {"$unset": {"metadata_local": 1, "media_local": 1}})
+            os.remove(file_path)
